Fixed Collision with collision types and did some optimization and fixed some warnings.

+++ AnimationComponent.cpp
@@ -22,7 +22,7 @@ const bool& AnimationComponent::isDone(const std::string key)
 	return this->animations[key]->isDone();
 }
 
-void AnimationComponent::addAnimation(const std::string key, float animation_timer, int start_frame_x, int start_frame_y, int frames_x, int frames_y, int width, int height)
+void AnimationComponent::addAnimation(const std::string key, float animation_timer, int start_frame_x, int start_frame_y, int frames_x, int frames_y, float width, float height)
 {
 	this->animations[key] = new Animation(this->sprite, this->textureSheet, animation_timer, start_frame_x, start_frame_y, frames_x, frames_y, width, height);
 }
