+++ AnimationComponent.h
@@ -108,7 +108,7 @@ public:
 
 	// Functions
 	const bool& isDone(const std::string key);
-	void addAnimation(const std::string key, float animation_timer, int start_frame_x, int start_frame_y, int frames_x, int frames_y, int width, int height);
+	void addAnimation(const std::string key, float animation_timer, int start_frame_x, int start_frame_y, int frames_x, int frames_y, float width, float height);
 	const bool& play(const std::string key, const float& dt, const bool priority = false);
 	const bool& play(const std::string key, const float& dt, const float& modifier, const float& modifier_max, const bool priority = false);
